requirements.txt
Pillow>=10.0.0
numpy>=1.24.0
# rembg>=3.0.0  # Uncomment when ready to use AI model
